1.перенес меню с кнопкой "поделиться" на экран деталей - логичнее там
